feat: 리포트 컴포넌트화 및 user에 따라 보이는 화면 수정
