+++ src/queries/keys/groups.key.ts
@@ -5,3 +5,7 @@ export const groupsQueryKeys = {
 export const groupsInviteQueryKeys = {
   inviteGroups: (id: number) => ['inviteGroups', id],
 };
+
+export const groupTasksQueryKeys = {
+  inviteGroups: (id: number) => ['groupTask', id],
+};
